fix: overflow & dmp_read_fifo

+++ src/dmp/mod.rs
@@ -398,8 +398,8 @@ pub struct DmpData {
     /// Host-calibrated accelerometer data
     pub host_calibrated_accel: Option<(i16, i16, i16)>,
 
-    /// Calibrated gyroscope data (raw - bias from DMP raw gyro stream)
-    pub calibrated_gyro: Option<(i16, i16, i16)>,
+    /// Calibrated gyroscope data (raw - bias); i32 because i16 - i16 can overflow
+    pub calibrated_gyro: Option<(i32, i32, i32)>,
 
     /// Calibrated gyroscope data from DMP (32-bit values)
     pub dmp_calibrated_gyro: Option<(i32, i32, i32)>,
